// src/app/dialogs/notifications-component/store/alert-configuration.types.ts
export enum AlertConfigurationActionTypes {
  LOAD_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Load alert configuration request",
  LOAD_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Load alert configuration Failure",
  LOAD_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Load alert configuration Success",

  LOAD_ARRHYTHMIA_ALERT_CONFIGURATION_REQUEST = "[ARRHYTHMIAALERTSCONFIGURATION] Arrhythmia load alert configuration request",
  LOAD_ARRHYTHMIA_ALERT_CONFIGURATION_FAILURE = "[ARRHYTHMIAALERTSCONFIGURATION] Arrhythmia load alert configuration Failure",
  LOAD_ARRHYTHMIA_ALERT_CONFIGURATION_SUCCESS = "[ARRHYTHMIAALERTSCONFIGURATION] Arrhythmia load alert configuration Success",

  UPDATE_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Update alert configuration request",
  UPDATE_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Update alert configuration Failure",
  UPDATE_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Update alert configuration Success",

  RESET_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Reset alert configuration request",
  RESET_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Reset alert configuration Failure",
  RESET_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Reset alert configuration Success",

  LOAD_DESTINATION_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Load Destination Alert Configuration Success",
  LOAD_DESTINATION_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Load Destination Alert Configuration Request",
  LOAD_DESTINATION_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Load Destination Alert Configuration Failure",

    CF_CONATCT_LIST_REQUEST = '[ALERTSCONFIGURATION] CF Contact list Request',
    CF_CONATCT_LIST_FAILURE = '[ALERTSCONFIGURATION] CF Contact list Failure',
    CF_CONATCT_LIST_SUCCESS = '[ALERTSCONFIGURATION] CF Contact list Success',

    LOAD_MISCELLANEOUS_SETTINGS_REQUEST = '[MISCELLANEOUSSETTINGS] Load miscellaneous settings request',
    LOAD_MISCELLANEOUS_SETTINGS_FAILURE = '[MISCELLANEOUSSETTINGS] Load miscellaneous settings Failure',
    LOAD_MISCELLANEOUS_SETTINGS_SUCCESS = '[MISCELLANEOUSSETTINGS] Load miscellaneous settings Success',

  UPDATE_MISCELLANEOUS_SETTINGS_REQUEST = "[MISCELLANEOUSSETTINGS] Update miscellaneous settings request",
  UPDATE_MISCELLANEOUS_SETTINGS_FAILURE = "[MISCELLANEOUSSETTINGS] Update miscellaneous settings Failure",
  UPDATE_MISCELLANEOUS_SETTINGS_SUCCESS = "[MISCELLANEOUSSETTINGS] Update miscellaneous settings Success",

  RESET_MISCELLANEOUS_SETTINGS_REQUEST = "[MISCELLANEOUSSETTINGS] Reset miscellaneous settings request",
  RESET_MISCELLANEOUS_SETTINGS_FAILURE = "[MISCELLANEOUSSETTINGS] Reset miscellaneous settings Failure",
  RESET_MISCELLANEOUS_SETTINGS_SUCCESS = "[MISCELLANEOUSSETTINGS] Reset miscellaneous settings Success",

  ALERT_HISTORY_REQUEST = "[ALERT HISTORY] Alert History Request",
  ALERT_HISTORY_SUCCESS = "[ALERT HISTORY] Alert History Success",
  ALERT_HISTORY_FAILURE = "[ALERT HISTORY] Alert History Failure",

  EVENT_LIST_REQUEST = "[EVENT LIST] Event List Request",
  EVENT_LIST_SUCCESS = "[EVENT LIST] Event List Success",
  EVENT_LIST_FAILURE = "[EVENT LIST] Event List Failure",

}